@@ -10,19 +10,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^19.0.0",
-    "@angular/common": "^19.0.0",
-    "@angular/compiler": "^19.0.0",
-    "@angular/core": "^19.0.0",
-    "@angular/forms": "^19.0.0",
-    "@angular/platform-browser": "^19.0.0",
-    "@angular/platform-browser-dynamic": "^19.0.0",
-    "@angular/router": "^19.0.0",
-    "@angular/service-worker": "^19.0.0",
+    "@angular/animations": "^19.2.0",
+    "@angular/common": "^19.2.0",
+    "@angular/compiler": "^19.2.0",
+    "@angular/core": "^19.2.0",
+    "@angular/fire": "^19.0.0",
+    "@angular/forms": "^19.2.0",
+    "@angular/platform-browser": "^19.2.0",
+    "@angular/platform-browser-dynamic": "^19.2.0",
+    "@angular/router": "^19.2.0",
+    "@angular/service-worker": "^19.2.0",
     "@tailwindcss/postcss": "^4.0.1",
-    "firebase": "^11.3.1",
     "postcss": "^8.5.1",
-    "rxfire": "^6.1.0",
     "rxjs": "~7.8.0",
     "tailwindcss": "^4.0.1",
     "tslib": "^2.3.0",
@@ -31,7 +30,7 @@
   "devDependencies": {
     "@angular-devkit/build-angular": "^19.0.2",
     "@angular/cli": "^19.0.2",
-    "@angular/compiler-cli": "^19.0.0",
+    "@angular/compiler-cli": "^19.2.0",
     "@types/jasmine": "~5.1.0",
     "jasmine-core": "~5.4.0",
     "karma": "~6.4.0",
